requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
httpx>=0.27.0
python-multipart>=0.0.12
# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4